pyproject.toml
[project]
name = "3-voice-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "twilio>=9.10.5",
    "websockets==11.0.3",
]
